mm: Move readahead limit outside of readahead, and advisory syscalls
Java workloads using the MappedByteBuffer library result in the fadvise()
and madvise() syscalls being used extensively. With readahead alterations
so as to limit the amount of readahead that is allowed, application
performance suffers in instances where small readahead is configured.

By moving this limit outside of the syscall codepaths, the syscalls are
able to advise an inordinately large amount of readahead when desired.
With a cap being imposed based on the half of NR_INACTIVE_FILE and
NR_FREE_PAGES. In essence, allowing performance tuning efforts to define
a small readahead limit, but then benefiting from large sequential
readahead values selectively.

Signed-off-by: Kyle Walker <kwalker@redhat.com>